From Mike Schwartz:
Some phone providers will clearly have a 1:1 line:did ratio. For these
providers, one form to contain both the line and which phone associated with
this line works.

However multiple DIDs can be routed to one trunk/line: for example, we might
route 10 DIDs to one IAX line. Based on the DID, the call is routed to a
different phone.

I created this configlet to enable you to route a DID from a certain line to
a certain phone.